schematic
The schematic of the programmer can be found from the schematic directory.
Removed the shift register of the data channel, so now data is written
directly to the memory chip's pins. This change makes verification of
the data written to the chip very easy.

To view the schematic, you'll need the arduino_pro_mini.lbr file (in the
schematic directory).