On the way to a working device
I built a (more or less) prototype version of the programmer on protoboard
today, and the code is now for testing (sketch.ino). Everything seems to work,
but a (maybe just Arduino-related) problem has occured: according to my
measurements, the logic LOW of the Arduino floats around 0.8..2.8 volts,
which means that the LOW level wont (always) be registered as LOW by the mem
chip, but rather as a HIGH -> pulldown resistors, maybe?

Also, the shiftreg writing system got simplified (I grew a brain and figured
out bitshifting..) -> better.